backend/requirements.txt
flask==2.0.2
flask-cors==3.0.10
pandas==1.3.4
psycopg2-binary==2.9.2
python-dotenv==0.19.2
sqlalchemy==1.4.27
pyjwt==2.3.0
exif==1.3.4
requests==2.26.0